+++ examples/ts/cli/map.ts
@@ -42,7 +42,6 @@ const client = new Client({
       debugInfo: true,
       emotions: true,
       multiModalActionPlanning: true,
-      logs: true,
       logsWarning: true,
       logsInfo: true,
       logsDebug: true,